Move: SystemAttack to the server
